dissolve borg.remote: server -> legacy.remote, cache helpers -> repository
borg.remote no longer fit its name: it held the legacy-only borg serve server
plus generic repository cache wrappers used by current repos. Split by purpose
and remove the module:

- Move RepositoryServer into borg.legacy.remote (it only serves legacy v1 ssh
  repositories). It reuses the exception classes (PathNotAllowed,
  InvalidRPCMethod, UnexpectedRPCDataFormatFromClient) and BORG_VERSION / MSGID
  constants already defined there; open() uses the module-level LegacyRepository.
  serve_cmd.py now imports RepositoryServer from ..legacy.remote.
- Move RepositoryNoCache and cache_if_remote into borg.repository (they wrap a
  Repository and are used by Archive.check and mount of current repos).
  archive.py and mount_cmds.py import them from ..repository now.
- Move the cache_if_remote tests into repository_test.py; delete remote_test.py.
- Delete src/borg/remote.py; fix the stale BUFSIZE comment in constants.py.

Pure relocation, no behavior change.
