+++ src/borg/repository.py
@@ -599,3 +599,58 @@ class Repository:
     def store_move(self, name, new_name=None, *, delete=False, undelete=False, deleted=False):
         self._lock_refresh()
         return self.store.move(name, new_name, delete=delete, undelete=undelete, deleted=deleted)
+
+
+class RepositoryNoCache:
+    """A Repository wrapper that passes through to the repository.
+
+    It applies an optional *transform* and provides a uniform context-manager API.
+
+    *transform* is a callable taking two arguments, key and raw repository data.
+    The return value is returned from get()/get_many(). By default, the raw
+    repository data is returned.
+    """
+
+    def __init__(self, repository, transform=None):
+        self.repository = repository
+        self.transform = transform or (lambda key, data: data)
+
+    def close(self):
+        pass
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_val, exc_tb):
+        self.close()
+
+    def get(self, key, read_data=True, raise_missing=True):
+        return next(self.get_many([key], read_data=read_data, raise_missing=raise_missing, cache=False))
+
+    def get_many(self, keys, read_data=True, raise_missing=True, cache=True):
+        for key, data in zip(keys, self.repository.get_many(keys, read_data=read_data, raise_missing=raise_missing)):
+            yield self.transform(key, data)
+
+    def log_instrumentation(self):
+        pass
+
+
+def cache_if_remote(repository, *, decrypted_cache=False, transform=None):
+    """
+    Return a RepositoryNoCache wrapping *repository*.
+
+    If *decrypted_cache* is a repo_objs object, then get and get_many will return a tuple
+    (csize, plaintext) instead of the actual data in the repository (the objects are
+    parsed/decrypted via the *transform* derived from it).
+    """
+    if decrypted_cache and transform:
+        raise ValueError("decrypted_cache and transform are incompatible")
+    elif decrypted_cache:
+        repo_objs = decrypted_cache
+
+        def transform(id_, data):
+            meta, decrypted = repo_objs.parse(id_, data, ro_type=ROBJ_DONTCARE)
+            csize = meta.get("csize", len(data))
+            return csize, decrypted
+
+    return RepositoryNoCache(repository, transform)
